+++ src/test/kotlin/io/csbroker/apiserver/controller/v1/admin/problem/AdminProblemControllerTest.kt
@@ -1,9 +1,8 @@
-package io.csbroker.apiserver.e2e.admin.problem
+package io.csbroker.apiserver.controller.v1.admin.problem
 
-import io.csbroker.apiserver.controller.v1.admin.problem.AdminProblemController
 import io.csbroker.apiserver.dto.problem.ProblemDeleteRequestDto
 import io.csbroker.apiserver.dto.problem.problemset.ProblemSetUpsertRequestDto
-import io.csbroker.apiserver.e2e.RestDocsTest
+import io.csbroker.apiserver.controller.RestDocsTest
 import io.csbroker.apiserver.service.problem.CommonProblemService
 import io.csbroker.apiserver.service.problem.ProblemSetService
 import io.mockk.every
