+++ src/test/kotlin/io/csbroker/apiserver/controller/v1/admin/problem/AdminShortProblemControllerTest.kt
@@ -1,10 +1,9 @@
-package io.csbroker.apiserver.e2e.admin.problem
+package io.csbroker.apiserver.controller.v1.admin.problem
 
-import io.csbroker.apiserver.controller.v1.admin.problem.AdminShortProblemController
 import io.csbroker.apiserver.dto.problem.shortproblem.ShortProblemResponseDto
 import io.csbroker.apiserver.dto.problem.shortproblem.ShortProblemSearchResponseDto
 import io.csbroker.apiserver.dto.problem.shortproblem.ShortProblemUpsertRequestDto
-import io.csbroker.apiserver.e2e.RestDocsTest
+import io.csbroker.apiserver.controller.RestDocsTest
 import io.csbroker.apiserver.service.problem.AdminShortProblemService
 import io.mockk.every
 import io.mockk.mockk
