+++ src/test/kotlin/io/csbroker/apiserver/controller/v2/problem/ProblemControllerV2Test.kt
@@ -1,4 +1,4 @@
-package io.csbroker.apiserver.e2e
+package io.csbroker.apiserver.controller.v2.problem
 
 import com.fasterxml.jackson.databind.ObjectMapper
 import io.csbroker.apiserver.auth.AuthTokenProvider
